pyproject.toml
[tool.poetry]
name = "yeg-dev-coe-scraper"
version = "0.1.0"
description = ""
authors = ["Dave Sutherland <dave@daveography.ca>"]

[tool.poetry.dependencies]
python = "^3.7"
selenium = "^3.141.0"
webdriver-manager = "^3.2.2"
SQLAlchemy = "^1.3.22"
PyYAML = "^5.3.1"
psycopg2 = "^2.8.6"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
flake8 = "^3.8.4"
pytest = "^6.2.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
